Procstat: collect connections and listener for each proc
Procstat modified to add a new metric "procstat_tcp" with to values
"conn" and "listen".
Those values will contain a comma separated value of the endpoints where
the proc is connecting to or listening (IPv4 and IPv6).

Local, virtual and docker interfaces are ignored.

If some proc is listening on 0.0.0.0, an endpoint for each of the
"public" (those not ignored as internal) IPv4 IPs is created.
If it is listening on :: (IPv6) an endpoint is created for each IPv4 and
IPv6.

For programs with one parent and several children, all listening in some
endpoint, only the parent process is taken into account. Child endpoints
are ignored.

Connections made to this cost (local port is one of the listening ports)
are ignored. To avoid having servers with thousands of connections.
We prefer to collect that info in the clients.

It is also added connection info (number of connections in each of the
TCP states) for each proc. Improving PR #5402

To gather tcp connections netlink is used, to avoid the cost of parsing
/proc/net/tcp(6), but /proc should be readed to get the mapping between
inodes and pids.